refactor : useSideMenu, useNotification, useNotificationSocket, useChatSocket, useCurrentUser
