+++ src/components/Chat/ChatRoomItem.tsx
@@ -2,7 +2,7 @@ import React, { FC, useCallback, useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { ChatRoom } from '../../typings';
 import { getFromNow } from '../../util/converDate';
-import { useSocket } from '../../hooks/useSocket';
+import { useChatSocket } from '../../hooks/useChatSocket';
 import { apis } from '../../apis';
 import { useQuery } from '@tanstack/react-query';
 
@@ -16,7 +16,7 @@ const ChatRoomItem: FC<Props> = ({ id, data }) => {
   const handleRoute = useCallback(() => {
     navigate(`/chat/${id}?title=${data.title}`);
   }, []);
-  const { chats } = useSocket(id);
+  const { chats } = useChatSocket(id);
   const { data: userList } = useQuery(['userList', data.roomId], () => apis.getRoomUserList(data.roomId));
 
   const lastMessage = useMemo(() => {
