+++ src/components/Common/Header.tsx
@@ -1,59 +1,13 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { toggleSideMenuShow } from '../../redux/features/commonSlice';
-import { RootState } from '../../redux/store';
-import { useNotification } from '../../hooks/useNotification';
-import dayjs from 'dayjs';
-import { useNavigate } from 'react-router-dom';
-import { apis } from '../../apis';
-import { useQuery, useQueryClient, useMutation } from '@tanstack/react-query';
-import { Notification } from '../../typings';
-import { toggleNotificationShow } from '../../redux/features/toggleSlice';
-import { setNotifications } from '../../redux/features/notificationSlice';
+import React from 'react';
+import useNotification from '../../hooks/useNotification';
+import useSideMenu from '../../hooks/useSideMenu';
 
 const alertPath = '/assets/images/alert.svg';
 const alertBasePath = '/assets/images/alert_base.svg';
 
 const Header = () => {
-  const { id, isLogin } = useSelector((state: RootState) => state.user);
-  const dispatch = useDispatch();
-  const navigate = useNavigate();
-  const queryClient = useQueryClient();
-  const notifications = useSelector((state: RootState) => state.notification.notifications);
-
-  const { refetch } = useQuery(['notifications'], apis.getNotifications, {
-    onSuccess: (data: Notification[]) => {
-      dispatch(setNotifications(data));
-    },
-    enabled: false,
-  });
-  const mutation = useMutation((id: number) => apis.postNotificationRead(id), {
-    onSuccess: () => {
-      queryClient.invalidateQueries(['notifications']);
-    },
-  });
-
-  useEffect(() => {
-    if (isLogin) {
-      refetch();
-    }
-  }, [isLogin]);
-
-  const toggleNotificationMenu = useCallback(() => {
-    dispatch(toggleNotificationShow());
-  }, []);
-
-  const toggleSideMenu = useCallback(() => {
-    dispatch(toggleSideMenuShow());
-  }, []);
-
-  const isNotificationExist = useMemo(() => {
-    if (notifications.length > 0) {
-      const res = notifications.some((v, idx) => !v.isread);
-      return res;
-    }
-    return false;
-  }, [notifications]);
+  const { toggleNotificationMenu, isNotificationExist } = useNotification();
+  const { toggleSideMenu } = useSideMenu();
 
   return (
     <header className='flex justify-end py-[14px]'>
